DBZ-8966 move to spi
