+++ debezium-core/src/main/java/io/debezium/processors/spi/PostProcessorFactory.java
@@ -4,11 +4,9 @@
  * Licensed under the Apache Software License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
  */
 
-package io.debezium.processors;
+package io.debezium.processors.spi;
 
 import java.util.List;
-
-import io.debezium.processors.spi.PostProcessor;
 
 public interface PostProcessorFactory {
